+++ app/src/main/AndroidManifest.xml
@@ -20,13 +20,13 @@
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".Views.main.MainActivity"
             android:label="@string/title_activity_the_main"
             android:theme="@style/AppTheme.LightStatusBar">
 
         </activity>
         <activity
-            android:name=".SplashScreenActivity"
+            android:name=".Views.SplashScreenActivity"
             android:screenOrientation="portrait"
             android:theme="@style/DefaultTheme">
             <intent-filter>
@@ -36,17 +36,18 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".StoryListingActivity" /> <!-- Do not edit this theme -->
-        <activity android:name=".StoryDetailActivity" />
-        <activity android:name=".SingleStoryActivity" />
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".ProfileRegisterActivity" />
-        <activity android:name=".ProfileActivity" />
-        <activity android:name=".OnBoardingActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".BookMark" />
-        <activity android:name=".AddStoryActivity" />
-        <activity android:name=".AddStoriesContentActivity" />
+        <activity android:name=".Views.StoryListingActivity" /> <!-- Do not edit this theme -->
+        <activity android:name=".Views.StoryDetailActivity" />
+        <activity android:name=".Views.SingleStoryActivity" />
+        <activity android:name=".Views.RegisterActivity" />
+        <activity android:name=".Views.ProfileRegisterActivity" />
+        <activity android:name=".Views.ProfileActivity" />
+        <activity android:name=".Views.OnBoardingActivity" />
+        <activity android:name=".Views.LoginActivity" />
+        <activity android:name=".Views.BookMark" />
+        <activity android:name=".Views.AddStoryActivity" />
+        <activity android:name=".Views.AddStoriesContentActivity" />
+        <activity android:name=".BaseActivity" />
 
         <meta-data
             android:name="com.facebook.sdk.ApplicationId"
