+++ app/src/main/java/com/project/android_kidstories/BaseActivity.java
@@ -1,0 +1,6 @@
+package com.project.android_kidstories;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class BaseActivity extends AppCompatActivity {
+}
