--- app/src/main/java/com/project/android_kidstories/ProfileRegisterActivity.java
@@ -1,14 +1,0 @@
-package com.project.android_kidstories;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class ProfileRegisterActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile_register);
-    }
-}
