+++ app/src/main/java/com/project/android_kidstories/Views/AddStoriesContentActivity.java
@@ -1,0 +1,16 @@
+package com.project.android_kidstories.Views;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.project.android_kidstories.R;
+
+public class AddStoriesContentActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_add_stories_content);
+    }
+}
