+++ app/src/main/java/com/project/android_kidstories/Views/BookMark.java
@@ -1,15 +1,16 @@
-package com.project.android_kidstories;
-
+package com.project.android_kidstories.Views;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class StoryListingActivity extends AppCompatActivity {
+import com.project.android_kidstories.R;
+
+public class BookMark extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_story_listing);
+        setContentView(R.layout.activity_book_mark);
     }
 }
