+++ app/src/main/java/com/project/android_kidstories/Views/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.project.android_kidstories;
+package com.project.android_kidstories.Views;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -27,6 +27,8 @@ import com.google.android.gms.auth.api.signin.GoogleSignInClient;
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
 import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.tasks.Task;
+import com.project.android_kidstories.Views.main.MainActivity;
+import com.project.android_kidstories.R;
 
 import java.util.Arrays;
