+++ app/src/main/java/com/project/android_kidstories/Views/ProfileRegisterActivity.java
@@ -1,4 +1,4 @@
-package com.project.android_kidstories;
+package com.project.android_kidstories.Views;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.project.android_kidstories.R;
 
-public class AddStoriesContentActivity extends AppCompatActivity {
+public class ProfileRegisterActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_stories_content);
+        setContentView(R.layout.activity_profile_register);
     }
 }
