+++ app/src/main/java/com/project/android_kidstories/Views/SingleStoryActivity.java
@@ -1,16 +1,14 @@
-package com.project.android_kidstories;
+package com.project.android_kidstories.Views;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
 import android.app.ProgressDialog;
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.project.android_kidstories.Model.Story;
+import com.project.android_kidstories.R;
 
 public class SingleStoryActivity extends AppCompatActivity {
