+++ app/src/main/java/com/project/android_kidstories/Views/StoryDetailActivity.java
@@ -1,9 +1,11 @@
-package com.project.android_kidstories;
+package com.project.android_kidstories.Views;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
 import android.os.Bundle;
+
+import com.project.android_kidstories.R;
 
 public class StoryDetailActivity extends AppCompatActivity {
