+++ app/src/main/java/com/project/android_kidstories/Views/StoryListingActivity.java
@@ -1,4 +1,5 @@
-package com.project.android_kidstories;
+package com.project.android_kidstories.Views;
+
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +7,11 @@ import android.os.Bundle;
 
 import com.project.android_kidstories.R;
 
-public class ProfileActivity extends AppCompatActivity {
+public class StoryListingActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile);
+        setContentView(R.layout.activity_story_listing);
     }
 }
