+++ app/src/main/java/com/project/android_kidstories/Views/main/ui/auth/LogOutFragment.java
@@ -1,4 +1,4 @@
-package com.project.android_kidstories.ui.auth;
+package com.project.android_kidstories.Views.main.ui.auth;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,14 +11,15 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProviders;
 
 import com.facebook.login.LoginManager;
-import com.project.android_kidstories.RegisterActivity;
+import com.project.android_kidstories.Views.RegisterActivity;
+
 
 public class LogOutFragment extends Fragment {
 
-    private LogOutViewModel logOutViewModel;
+    private com.project.android_kidstories.ui.auth.LogOutViewModel logOutViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        logOutViewModel = ViewModelProviders.of(this).get(LogOutViewModel.class);
+        logOutViewModel = ViewModelProviders.of(this).get(com.project.android_kidstories.ui.auth.LogOutViewModel.class);
         // Logout of facebook
         if (LoginManager.getInstance() != null) {
             LoginManager.getInstance().logOut();
