+++ app/src/main/java/com/project/android_kidstories/Views/main/ui/home/Fragments/BaseFragment.java
