+++ app/src/main/res/layout/activity_add_stories_content.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddStoriesContentActivity">
+    tools:context=".Views.AddStoriesContentActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
