+++ app/src/main/res/layout/activity_on_boarding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnBoardingActivity"
+    tools:context=".Views.OnBoardingActivity"
     android:background="#ffffff">
 
     <androidx.viewpager.widget.ViewPager
